+++ src/scripts/dashboard/chart/bar.test.js
@@ -28,9 +28,9 @@ describe("scripts - dashboard: bar chart", () => {
                 datasets: [
                     {
                         data: [1, 2, 3],
-                        backgroundColor: ["rgba(TEST, 0.5)", "rgba(TEST, 0.5)", "rgba(TEST, 0.5)"],
-                        borderColor: ["rgba(TEST, 0.9)", "rgba(TEST, 0.9)", "rgba(TEST, 0.9)"],
-                        hoverBackgroundColor: ["rgba(TEST, 0.7)", "rgba(TEST, 0.7)", "rgba(TEST, 0.7)"],
+                        backgroundColor: ["rgba(TEST, 1)", "rgba(TEST, 1)", "rgba(TEST, 1)"],
+                        borderColor: ["rgba(TEST, 1)", "rgba(TEST, 1)", "rgba(TEST, 1)"],
+                        hoverBackgroundColor: ["rgba(TEST, 1)", "rgba(TEST, 1)", "rgba(TEST, 1)"],
                         hoverBorderColor: ["rgba(TEST, 1)", "rgba(TEST, 1)", "rgba(TEST, 1)"],
                         borderWidth: 1,
                         hoverBorderWidth: 2
@@ -69,9 +69,9 @@ describe("scripts - dashboard: bar chart", () => {
                 datasets: [
                     {
                         data: [1, 2, 3],
-                        backgroundColor: ["rgba(TEST, 0.5)", "rgba(TEST, 0.5)", "rgba(TEST, 0.5)"],
-                        borderColor: ["rgba(TEST, 0.9)", "rgba(TEST, 0.9)", "rgba(TEST, 0.9)"],
-                        hoverBackgroundColor: ["rgba(TEST, 0.7)", "rgba(TEST, 0.7)", "rgba(TEST, 0.7)"],
+                        backgroundColor: ["rgba(TEST, 1)", "rgba(TEST, 1)", "rgba(TEST, 1)"],
+                        borderColor: ["rgba(TEST, 1)", "rgba(TEST, 1)", "rgba(TEST, 1)"],
+                        hoverBackgroundColor: ["rgba(TEST, 1)", "rgba(TEST, 1)", "rgba(TEST, 1)"],
                         hoverBorderColor: ["rgba(TEST, 1)", "rgba(TEST, 1)", "rgba(TEST, 1)"],
                         borderWidth: 1,
                         hoverBorderWidth: 2
@@ -79,16 +79,16 @@ describe("scripts - dashboard: bar chart", () => {
                     {
                         type: "line",
                         data: [1, 2, 3],
-                        backgroundColor: ["rgba(TEST2, 0.2)", "rgba(TEST2, 0.2)", "rgba(TEST2, 0.2)"],
-                        borderColor: ["rgba(TEST2, 0.7)", "rgba(TEST2, 0.7)", "rgba(TEST2, 0.7)"],
-                        pointBackgroundColor: ["rgba(TEST2, 0.9)", "rgba(TEST2, 0.9)", "rgba(TEST2, 0.9)"],
-                        pointBorderColor: ["rgba(TEST2, 0.9)", "rgba(TEST2, 0.9)", "rgba(TEST2, 0.9)"],
-                        pointHoverBorderColor: ["rgba(TEST2, 1)", "rgba(TEST2, 1)", "rgba(TEST2, 1)"],
-                        pointHoverBackgroundColor: ["rgba(TEST2, 1)", "rgba(TEST2, 1)", "rgba(TEST2, 1)"],
+                        backgroundColor: "rgba(TEST2, 1)",
+                        borderColor: "rgba(TEST2, 1)",
+                        pointBackgroundColor: "rgba(TEST2, 1)",
+                        pointBorderColor: "rgba(TEST2, 1)",
+                        pointHoverBorderColor: "rgba(TEST2, 1)",
+                        pointHoverBackgroundColor: "rgba(TEST2, 1)",
                         borderWidth: 4,
                         hoverBorderWidth: 2,
                         pointBorderWidth: 2,
-                        pointRadius: 4,
+                        pointRadius: 3,
                         pointHoverRadius: 6
                     }
                 ]